requirements.txt
asgiref==3.3.1
Django==3.1.3
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==1.10.0
django-crontab==0.7.1
gunicorn==20.1.0
Pillow==8.0.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
